@@ -16,7 +16,6 @@
 
 import { inject, injectable } from 'inversify';
 import {
-    ConfigLoader,
     Scraper,
     TASK_EXCHANGE,
     TASK_QUEUE,
@@ -31,30 +30,31 @@ import { logger } from '../utils/logger-factory';
 import { MQMessage, RabbitMQService, TYPES } from '@irohalab/mira-shared';
 import { promisify } from 'util';
 import Timeout = NodeJS.Timeout;
+import { ConfigManager } from '../utils/config-manager';
 
 const sleep = promisify(setTimeout) as (t: number) => Promise<unknown>;
 
 @injectable()
 export class TaskOrchestra {
-    private _timerId: Timeout;
+    private _checkMainTaskTimerId: Timeout;
+    private _checkFailedTaskTimerId: Timeout;
     private _lastMainTaskExeTime  = 0;
     private _scraper: Scraper;
     private _lastExeTime = 0;
 
     private readonly _exchangeName: string;
 
-    constructor(@inject(TYPES_IDX.ConfigLoader) private _config: ConfigLoader,
+    constructor(@inject(TYPES.ConfigManager) private _config: ConfigManager,
                 @inject(TYPES_IDX.TaskTimingFactory) private _taskTimingFactory: (interval: number) => number,
                 @inject(TYPES.RabbitMQService) private _mqService: RabbitMQService,
                 @inject(TYPES_IDX.TaskStorage) private _taskStore: TaskQueue,
                 @inject(TYPES_IDX.ThrottleStore) private _throttleStore: ThrottleStore) {
-        this._exchangeName = `${TASK_EXCHANGE}_${this._config.mode}`;
+        this._exchangeName = `${TASK_EXCHANGE}_${this._config.getMode()}`;
     }
 
     public async start(scraper: Scraper): Promise<void> {
-        let actualInterval = this._taskTimingFactory(this._config.minInterval);
+        let actualInterval = this._taskTimingFactory(this._config.getMinInterval());
         this._scraper = scraper;
-        await this.checkMainTask(actualInterval);
         await this._mqService.initPublisher(this._exchangeName, 'direct', TASK_ROUTING_KEY);
         await this._mqService.initConsumer(this._exchangeName, 'direct', TASK_QUEUE, TASK_ROUTING_KEY);
         await this._mqService.consume(TASK_QUEUE, async (msg: MQMessage) => {
@@ -64,15 +64,17 @@ export class TaskOrchestra {
                 await sleep(2000);
                 return false;
             }
-            let result = await this._scraper.executeTask(task);
             this._lastExeTime = Date.now();
+            let result = await this._scraper.executeTask(task);
             if (result === TaskStatus.NeedRetry) {
-                return false;
+                await this._taskStore.offerFailedTask(task);
             } else if (result === TaskStatus.Success && task.type === TaskType.MAIN) {
                 this._lastMainTaskExeTime = Date.now();
             }
             return true;
         });
+        await this.checkMainTask(actualInterval);
+        await this.checkFailedTask();
     }
 
     public async queue(task: Task): Promise<void> {
@@ -80,25 +82,42 @@ export class TaskOrchestra {
     }
 
     public stop() {
-        clearTimeout(this._timerId);
+        clearTimeout(this._checkMainTaskTimerId);
+        clearTimeout(this._checkFailedTaskTimerId);
     }
 
     private async checkMainTask(actualInterval: number): Promise<void> {
-        const lastMainTaskTime = this._throttleStore.getLastMainTaskTime();
+        const lastMainTaskTime = await this._throttleStore.getLastMainTaskTime();
         const currentTime = Date.now();
         const delta = currentTime - lastMainTaskTime;
-        if (delta >= this._config.minCheckInterval && currentTime >= this._lastExeTime + actualInterval) {
+        if (delta >= this._config.getMinCheckInterval() && currentTime >= this._lastExeTime + actualInterval) {
             this._lastExeTime = currentTime;
             await this._throttleStore.setLastMainTaskTime();
             await this._scraper.executeTask(new CommonTask(TaskType.MAIN));
         }
-        this._timerId = setTimeout(async () => {
+        this._checkMainTaskTimerId = setTimeout(async () => {
             await this.checkMainTask(actualInterval);
         }, actualInterval);
     }
 
+    private async checkFailedTask(): Promise<void> {
+        const lastFailedTaskTime = await this._throttleStore.getLastFailedTaskTime();
+        const currentTime = Date.now();
+        const delta = currentTime - lastFailedTaskTime;
+        if (delta >= this._config.getMinFailedTaskCheckInterval()) {
+            await this._throttleStore.setLastFailedTaskTime();
+            while (await this.pickFailedTask()) {
+                await sleep(1000); // some small interval to queue task.
+            }
+        }
+        const checkInterval = this._taskTimingFactory(this._config.getMinFailedTaskCheckInterval() / 2);
+        this._checkFailedTaskTimerId = setTimeout(async () => {
+            await this.checkFailedTask();
+        }, checkInterval);
+    }
+
     private pick(): void {
-        let actualInterval = this._taskTimingFactory(this._config.minInterval);
+        let actualInterval = this._taskTimingFactory(this._config.getMinInterval());
         this.pickTask()
             .then((hasTask) => {
                 /* if task queue is empty, try failed task */
@@ -109,15 +128,15 @@ export class TaskOrchestra {
             })
             .then((hasSomeTaskExecuted) => {
                 /* Need to ensure the list page is checked regularly */
-                if (Date.now() - this._lastMainTaskExeTime > this._config.minCheckInterval) {
+                if (Date.now() - this._lastMainTaskExeTime > this._config.getMinCheckInterval()) {
                     // force queue a MainTask
                     logger.info('force_queue', {
                         last_main_task_execute_time: this._lastMainTaskExeTime,
-                        min_check_interval: this._config.minCheckInterval
+                        min_check_interval: this._config.getMinCheckInterval()
                     });
                     this.queue(new CommonTask(TaskType.MAIN))
                         .then(() => {
-                            this._timerId = setTimeout(() => {
+                            this._checkMainTaskTimerId = setTimeout(() => {
                                 this.pick();
                             }, actualInterval);
                         });
@@ -125,7 +144,7 @@ export class TaskOrchestra {
                 }
                 /* Either task or failed task has been executed, we will try to pick again */
                 if (hasSomeTaskExecuted) {
-                    this._timerId = setTimeout(() => {
+                    this._checkMainTaskTimerId = setTimeout(() => {
                         this.pick();
                     }, actualInterval);
                 } else {
@@ -133,9 +152,9 @@ export class TaskOrchestra {
                     this.queue(new CommonTask(TaskType.MAIN))
                         .then(() => {
                             let offset = Date.now() - this._lastMainTaskExeTime;
-                            this._timerId = setTimeout(() => {
+                            this._checkMainTaskTimerId = setTimeout(() => {
                                 this.pick();
-                            }, Math.max(this._config.minCheckInterval - offset, actualInterval));
+                            }, Math.max(this._config.getMinCheckInterval() - offset, actualInterval));
                         });
                 }
             });
@@ -166,17 +185,14 @@ export class TaskOrchestra {
     private async pickFailedTask(): Promise<boolean> {
         if (await this._taskStore.hasFailedTask()) {
             let task = await this._taskStore.pollFailedTask();
-            if (task.retryCount > this._config.maxRetryCount) {
+            if (task.retryCount > this._config.getMaxRetryCount()) {
                 // drop task
                 logger.warn('drop_task', {
                     task
                 });
-                return false;
+                return this.pickFailedTask();
             }
-            let result = await this._scraper.executeTask(task);
-            if (result === TaskStatus.NeedRetry) {
-                await this._taskStore.offerFailedTask(task);
-            }
+            await this.queue(task);
             return true;
         }
         return false;
