@@ -17,13 +17,13 @@
 import { createServer, Server } from 'http';
 import { Container } from 'inversify';
 import { InversifyExpressServer } from 'inversify-express-utils';
-import { ConfigLoader } from './TYPES_IDX';
 import { logger } from './utils/logger-factory';
+import { ConfigManager } from './utils/config-manager';
 
 export class RESTServer {
     private _server: InversifyExpressServer;
-    private _config: ConfigLoader;
-    constructor(container: Container, config: ConfigLoader) {
+    private _config: ConfigManager;
+    constructor(container: Container, config: ConfigManager) {
         this._config = config;
         this._server = new InversifyExpressServer(container);
     }
@@ -35,6 +35,6 @@ export class RESTServer {
             logger.info('stopping REST server...');
             httpServer.close();
         });
-        return httpServer.listen(this._config.serverPort, this._config.serverHost);
+        return httpServer.listen(this._config.getServerPort(), this._config.getServerPort());
     }
 }
