@@ -20,27 +20,29 @@ import { FakeConfigManager } from '../test/fake-config';
 import { FakeResource, FakeScraper, MIN_INTERVAL } from '../test/fake-scraper';
 import { InMemoryTaskStore } from '../test/in-memory-task-store';
 import { MockTaskTiming } from '../test/mock-task-timing';
-import { ConfigLoader, Scraper, TaskQueue, TYPES_IDX } from '../TYPES_IDX';
+import { Scraper, TaskQueue, TYPES_IDX } from '../TYPES_IDX';
 import { TaskOrchestra } from './task-orchestra';
+import { ConfigManager } from '../utils/config-manager';
+import { TYPES } from '@irohalab/mira-shared';
 
 @TestFixture('TaskOrchestra test spec')
 export class TaskOrchestraSpec {
     private _container: Container;
-    private _config: ConfigLoader;
+    private _config: ConfigManager;
     private _scraper: FakeScraper;
 
     @SetupFixture
     public setupFixture() {
         this._container = new Container();
-        this._container.bind<ConfigLoader>(TYPES_IDX.ConfigLoader).to(FakeConfigManager).inSingletonScope();
+        this._container.bind<ConfigManager>(TYPES.ConfigManager).to(FakeConfigManager).inSingletonScope();
         this._container.bind<interfaces.Factory<number>>(TYPES_IDX.TaskTimingFactory).toFactory<number>(MockTaskTiming);
         this._container.bind<TaskQueue>(TYPES_IDX.TaskStorage).to(InMemoryTaskStore).inSingletonScope();
         this._container.bind<TaskOrchestra>(TaskOrchestra).toSelf();
         this._container.bind<Scraper>(TYPES_IDX.Scraper).to(FakeScraper).inTransientScope();
-        this._config = this._container.get<ConfigLoader>(TYPES_IDX.ConfigLoader);
-        this._config.load();
-        this._config.minCheckInterval = MIN_INTERVAL * 3;
-        this._config.minInterval = MIN_INTERVAL;
+        this._config = this._container.get<ConfigManager>(TYPES.ConfigManager);
+        this._config.prepare();
+        (this._config as unknown as FakeConfigManager).minCheckInterval = MIN_INTERVAL * 3;
+        (this._config as unknown as FakeConfigManager).minInterval = MIN_INTERVAL;
     }
 
     @Setup
